requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.1.31
distro==1.9.0
Django==4.2.19
django-environ==0.11.2
djangorestframework==3.15.2
exceptiongroup==1.2.2
groq==0.18.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
packaging==24.2
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
slack_sdk==3.34.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
celery>=5.3.0
redis>=5.0.0
boto3>=1.26.0
requests